# doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rec an Dec API",
    "description": "API para la aplicacion de gestion de egresos e ingresos Rec an Dec",
    "contact": {
      "name": "Igna Garcia",
      "url": "https://ignagarcia.vercel.app/",
      "email": "gnachoxp@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3100"
    }
  ],
...
  "paths": {
...
    "/tags/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "example": "any"
                  },
                  "alias": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
...
}